requirements.txt
langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
